Fixed minor type-conversion warnings
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@1026 3db68df2-c116-0410-a063-a993310a9797

+++ source/ps/Interact.cpp
@@ -418,8 +418,8 @@ void CSelectedEntities::contextOrder( bool pushQueue )
 	{
 		CGotoCommand *msg=new CGotoCommand();
 		msg->m_Entity=m_selected[0]->me;
-		msg->m_TargetX=g_Mouseover.m_worldposition.x;
-		msg->m_TargetY=g_Mouseover.m_worldposition.y;
+		msg->m_TargetX=(u32)g_Mouseover.m_worldposition.x;
+		msg->m_TargetY=(u32)g_Mouseover.m_worldposition.y;
 		g_Game->GetSimulation()->QueueLocalCommand(msg);
 		break;
 	}
